requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.39.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
chex==0.1.82
click==8.1.7
dataclasses-json==0.5.14
docopt==0.6.2
etils==1.4.1
filelock==3.12.2
flax==0.7.0
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.2
langchain==0.0.274
langsmith==0.0.26
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
opt-einsum==3.3.0
optax==0.1.4
orbax-checkpoint==0.3.5
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.10.0
protobuf==4.24.1
psutil==5.9.5
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
safetensors==0.3.3
scipy==1.11.2
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
tensorstore==0.1.41
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
